Update metadata suffix and headers; modify page title composition
This commit revises the metadata file to change the arrangement of the author's name in the suffix. It also alters the bootstrap settings of the navigation menu in header.njk to center-align in all screen sizes. Furthermore, it rewrites the method of composing the page's title in metagen.njk based on different permalink conditions, and the Atom and JSON feed links are removed.